+++ collection-service/__brick__/{{collectionName.snakeCase()}}_service.dart
@@ -1,0 +1,22 @@
+class {{collectionName.pascalCase()}}Service extends TurboCollectionService<{{collectionObjectName.pascalCase()}}, {{collectionName.pascalCase()}}Api> {
+  {{collectionName.pascalCase()}}Service() : super(api: {{collectionName.pascalCase()}}Api.locate);
+
+  // 📍 LOCATOR ------------------------------------------------------------------------------- \\
+
+  static {{collectionName.pascalCase()}}Service get locate => GetIt.I.get();
+  static void registerLazySingleton() => GetIt.I.registerLazySingleton({{collectionName.pascalCase()}}Service.new);
+
+  // 🧩 DEPENDENCIES -------------------------------------------------------------------------- \\
+  // 🎬 INIT & DISPOSE ------------------------------------------------------------------------ \\
+  // 👂 LISTENERS ----------------------------------------------------------------------------- \\
+  // ⚡️ OVERRIDES ----------------------------------------------------------------------------- \\
+
+  @override
+  FutureOr<Stream<List<{{collectionObjectName.pascalCase()}}>>> Function(User user) get stream => throw UnimplementedError();
+
+// 🎩 STATE --------------------------------------------------------------------------------- \\
+// 🛠 UTIL ---------------------------------------------------------------------------------- \\
+// 🧲 FETCHERS ------------------------------------------------------------------------------ \\
+// 🏗️ HELPERS ------------------------------------------------------------------------------- \\
+// 🪄 MUTATORS ------------------------------------------------------------------------------ \\
+}