+++ core/__brick__/lib/core/constants/k_lottie.dart
@@ -1,8 +1,0 @@
-const kLottie = ConstLottie._();
-
-class ConstLottie {
-  const ConstLottie._();
-
-  static const _confetti = 'assets/lottie/confetti.json';
-  String get confetti => _confetti;
-}