+++ core/__brick__/lib/core/enums/animation_timing.dart
@@ -1,8 +1,0 @@
-enum AnimationTiming {
-  before,
-  after,
-  ;
-
-  bool get isBefore => this == AnimationTiming.before;
-  bool get isAfter => this == AnimationTiming.after;
-}