+++ core/__brick__/lib/core/enums/device_type.dart
@@ -1,9 +1,0 @@
-enum DeviceType {
-  mobile,
-  tablet,
-  web;
-
-  bool get isMobile => this == DeviceType.mobile;
-  bool get isTabletDesktop => this == DeviceType.tablet;
-  bool get isWeb => this == DeviceType.web;
-}