+++ core/__brick__/lib/core/enums/form_field_type.dart
@@ -1,8 +1,0 @@
-enum FormFieldType {
-  text,
-  dropDown,
-  checkbox,
-  counter,
-  chipWrap,
-  dateTimePicker;
-}