+++ core/__brick__/lib/core/globals/g_env.dart
@@ -1,7 +1,0 @@
-import 'package:{{packageName.snakeCase()}}/core/enums/environment.dart';
-import 'package:{{packageName.snakeCase()}}/core/enums/supported_platform.dart';
-
-SupportedPlatform get gPlatform => SupportedPlatform.current;
-EnvironmentType get gEnvironment => Environment.current;
-bool get gIsEmulators => Environment.isEmulators;
-bool get gIsProd => Environment.isProd;