+++ core/__brick__/lib/core/globals/g_is_busy.dart
@@ -1,8 +1,0 @@
-/// Global busy state indicator
-bool gIsBusy = false;
-
-/// Set the global busy state
-void gSetBusy() => gIsBusy = true;
-
-/// Set the global idle state
-void gSetIdle() => gIsBusy = false;