+++ core/__brick__/lib/core/globals/g_now.dart
@@ -1,2 +1,0 @@
-/// Returns the current [DateTime] using [DateTime.now].
-DateTime get gNow => DateTime.now();