+++ core/__brick__/lib/core/globals/g_strings.dart
@@ -1,5 +1,0 @@
-import '../../../core/strings/gen/l10n.dart';
-
-/// Returns the current instance of [Strings] for localization.
-Strings get gStrings => Strings.current;
-String get gUnknown => Strings.current.unknown;